+++ apps/frontend/src/libs/components/components.ts
@@ -1,5 +1,7 @@
 export { App } from "./app/app.js";
 export { Button } from "./button/button.js";
+export { Checkbox } from "./checkbox/checkbox.js";
+export { Content } from "./content/content.js";
 export { Courses } from "./courses/courses.js";
 export { Image } from "./image/image.js";
 export { Input } from "./input/input.js";