+++ apps/frontend/src/pages/course-description/libs/components/course-details/libs/maps/maps.ts
@@ -1,0 +1,1 @@
+export { tabToReadable } from "./tab-to-readable.map.js";