+++ packages/shared/src/index.ts
@@ -18,6 +18,7 @@ export {
 	configureString,
 	getDifferenceInHours,
 	getFormattedDate,
+	getSanitizedHtml,
 	getSizeInBytes,
 	initDebounce,
 } from "./libs/helpers/helpers.js";
@@ -70,7 +71,9 @@ export {
 	CourseSectionError,
 	type CourseSectionGetAllRequestDto,
 	type CourseSectionGetAllResponseDto,
+	type CourseSectionWithStatusDto,
 	CourseSectionsApiPath,
+	courseSectionGetAllQueryValidationSchema,
 	courseSectionIdParameterValidationSchema,
 } from "./modules/course-sections/course-sections.js";
 export {
@@ -101,6 +104,20 @@ export {
 	addFriendValidationSchema,
 	friendIdParameterValidationSchema,
 } from "./modules/friends/friends.js";
+export {
+	SectionStatus,
+	type SectionStatusAddRequestDto,
+	SectionStatusError,
+	type SectionStatusGetAllRequestDto,
+	type SectionStatusGetAllResponseDto,
+	type SectionStatusResponseDto,
+	type SectionStatusUpdateRequestDto,
+	SectionStatusesApiPath,
+	sectionStatusCreateBodyValidationSchema,
+	sectionStatusGetAllQueryValidationSchema,
+	sectionStatusUpdateBodyValidationSchema,
+	sectionStatusUpdateQueryValidationSchema,
+} from "./modules/section-statuses/section-statuses.js";
 export {
 	UserCoursesApiPath,
 	userCourseGetAllQueryValidationSchema,