+++ packages/shared/src/libs/helpers/get-sanitized-html/get-sanitized-html.helper.ts
@@ -1,0 +1,7 @@
+import sanitizeHtml from "sanitize-html";
+
+const getSanitizedHtml = (content: string): string => {
+	return sanitizeHtml(content);
+};
+
+export { getSanitizedHtml };