+++ packages/shared/src/modules/course-sections/libs/validation-schemas/course-section-get-all-query.validation-schema.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+const courseSectionGetAllQuery = z
+	.object({
+		courseId: z.coerce.number(),
+	})
+	.required();
+
+export { courseSectionGetAllQuery };